+++ openmetadata-ui/src/main/resources/ui/src/pages/EntityVersionPage/EntityVersionPage.component.tsx
@@ -14,6 +14,7 @@
 import { isEmpty } from 'lodash';
 import {
   FunctionComponent,
+  Suspense,
   useCallback,
   useEffect,
   useMemo,
@@ -21,26 +22,9 @@ import {
 } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useNavigate } from 'react-router-dom';
-import APIEndpointVersion from '../../components/APIEndpoint/APIEndpointVersion/APIEndpointVersion';
-import ChartVersion from '../../components/Chart/ChartVersion/ChartVersion.component';
 import ErrorPlaceHolder from '../../components/common/ErrorWithPlaceholder/ErrorPlaceHolder';
 import Loader from '../../components/common/Loader/Loader';
-import ContainerVersion from '../../components/Container/ContainerVersion/ContainerVersion.component';
-import DashboardVersion from '../../components/Dashboard/DashboardVersion/DashboardVersion.component';
-import DataModelVersion from '../../components/Dashboard/DataModel/DataModelVersion/DataModelVersion.component';
-import StoredProcedureVersion from '../../components/Database/StoredProcedureVersion/StoredProcedureVersion.component';
-import TableVersion from '../../components/Database/TableVersion/TableVersion.component';
-import DataProductsPage from '../../components/DataProducts/DataProductsPage/DataProductsPage.component';
-import DirectoryVersion from '../../components/DriveService/Directory/DirectoryVersion/DirectoryVersion';
-import FileVersion from '../../components/DriveService/File/FileVersion/FileVersion';
-import SpreadsheetVersion from '../../components/DriveService/Spreadsheet/SpreadsheetVersion/SpreadsheetVersion';
-import WorksheetVersion from '../../components/DriveService/Worksheet/WorksheetVersion/WorksheetVersion';
-import MetricVersion from '../../components/Metric/MetricVersion/MetricVersion';
-import MlModelVersion from '../../components/MlModel/MlModelVersion/MlModelVersion.component';
 import PageLayoutV1 from '../../components/PageLayoutV1/PageLayoutV1';
-import PipelineVersion from '../../components/Pipeline/PipelineVersion/PipelineVersion.component';
-import SearchIndexVersion from '../../components/SearchIndexVersion/SearchIndexVersion';
-import TopicVersion from '../../components/Topic/TopicVersion/TopicVersion.component';
 import { usePermissionProvider } from '../../context/PermissionProvider/PermissionProvider';
 import {
   OperationPermission,
@@ -135,13 +119,11 @@ import {
 } from '../../rest/topicsAPI';
 import entityUtilClassBase from '../../utils/EntityUtilClassBase';
 import { getEntityBreadcrumbs, getEntityName } from '../../utils/EntityUtils';
+import entityVersionClassBase from '../../utils/EntityVersionClassBase';
 import { DEFAULT_ENTITY_PERMISSION } from '../../utils/PermissionsUtils';
 import { getEntityDetailsPath, getVersionPath } from '../../utils/RouterUtils';
 import { getTierTags } from '../../utils/TableUtils';
 import { useRequiredParams } from '../../utils/useRequiredParams';
-import APICollectionVersionPage from '../APICollectionPage/APICollectionVersionPage';
-import DatabaseSchemaVersionPage from '../DatabaseSchemaVersionPage/DatabaseSchemaVersionPage';
-import DatabaseVersionPage from '../DatabaseVersionPage/DatabaseVersionPage';
 import './EntityVersionPage.less';
 
 export type VersionData =
@@ -610,336 +592,441 @@ const EntityVersionPage: FunctionComponent = () => {
       );
     }
 
-    let VersionPage = null;
+    const TableVersion = entityVersionClassBase.getEntityVersionComponent(
+      EntityType.TABLE
+    );
+    const TopicVersion = entityVersionClassBase.getEntityVersionComponent(
+      EntityType.TOPIC
+    );
+    const DashboardVersion = entityVersionClassBase.getEntityVersionComponent(
+      EntityType.DASHBOARD
+    );
+    const PipelineVersion = entityVersionClassBase.getEntityVersionComponent(
+      EntityType.PIPELINE
+    );
+    const MlModelVersion = entityVersionClassBase.getEntityVersionComponent(
+      EntityType.MLMODEL
+    );
+    const ContainerVersion = entityVersionClassBase.getEntityVersionComponent(
+      EntityType.CONTAINER
+    );
+    const SearchIndexVersion = entityVersionClassBase.getEntityVersionComponent(
+      EntityType.SEARCH_INDEX
+    );
+    const DataModelVersion = entityVersionClassBase.getEntityVersionComponent(
+      EntityType.DASHBOARD_DATA_MODEL
+    );
+    const StoredProcedureVersion =
+      entityVersionClassBase.getEntityVersionComponent(
+        EntityType.STORED_PROCEDURE
+      );
+    const APIEndpointVersion = entityVersionClassBase.getEntityVersionComponent(
+      EntityType.API_ENDPOINT
+    );
+    const MetricVersion = entityVersionClassBase.getEntityVersionComponent(
+      EntityType.METRIC
+    );
+    const ChartVersion = entityVersionClassBase.getEntityVersionComponent(
+      EntityType.CHART
+    );
+    const DirectoryVersion = entityVersionClassBase.getEntityVersionComponent(
+      EntityType.DIRECTORY
+    );
+    const FileVersion = entityVersionClassBase.getEntityVersionComponent(
+      EntityType.FILE
+    );
+    const SpreadsheetVersion = entityVersionClassBase.getEntityVersionComponent(
+      EntityType.SPREADSHEET
+    );
+    const WorksheetVersion = entityVersionClassBase.getEntityVersionComponent(
+      EntityType.WORKSHEET
+    );
+    const DatabaseVersionPage =
+      entityVersionClassBase.getEntityVersionComponent(EntityType.DATABASE);
+    const DatabaseSchemaVersionPage =
+      entityVersionClassBase.getEntityVersionComponent(
+        EntityType.DATABASE_SCHEMA
+      );
+    const DataProductsPage = entityVersionClassBase.getEntityVersionComponent(
+      EntityType.DATA_PRODUCT
+    );
+    const APICollectionVersionPage =
+      entityVersionClassBase.getEntityVersionComponent(
+        EntityType.API_COLLECTION
+      );
+
+    const wrapSuspense = (node: JSX.Element) => (
+      <Suspense fallback={<Loader />}>{node}</Suspense>
+    );
 
     switch (entityType) {
       case EntityType.TABLE: {
-        return (
-          <TableVersion
-            backHandler={backHandler}
-            currentVersionData={currentVersionData as Table}
-            dataProducts={currentVersionData.dataProducts}
-            deleted={currentVersionData.deleted}
-            domains={domains}
-            entityPermissions={entityPermissions}
-            isVersionLoading={isVersionLoading}
-            owners={owners}
-            slashedTableName={slashedEntityName}
-            tier={tier as TagLabel}
-            version={version}
-            versionHandler={versionHandler}
-            versionList={versionList}
-          />
-        );
+        return TableVersion
+          ? wrapSuspense(
+              <TableVersion
+                backHandler={backHandler}
+                currentVersionData={currentVersionData as Table}
+                dataProducts={currentVersionData.dataProducts}
+                deleted={currentVersionData.deleted}
+                domains={domains}
+                entityPermissions={entityPermissions}
+                isVersionLoading={isVersionLoading}
+                owners={owners}
+                slashedTableName={slashedEntityName}
+                tier={tier as TagLabel}
+                version={version}
+                versionHandler={versionHandler}
+                versionList={versionList}
+              />
+            )
+          : null;
       }
       case EntityType.TOPIC: {
-        return (
-          <TopicVersion
-            backHandler={backHandler}
-            currentVersionData={currentVersionData as Topic}
-            dataProducts={currentVersionData.dataProducts}
-            deleted={currentVersionData.deleted}
-            domains={domains}
-            entityPermissions={entityPermissions}
-            isVersionLoading={isVersionLoading}
-            owners={owners}
-            slashedTopicName={slashedEntityName}
-            tier={tier as TagLabel}
-            version={version}
-            versionHandler={versionHandler}
-            versionList={versionList}
-          />
-        );
+        return TopicVersion
+          ? wrapSuspense(
+              <TopicVersion
+                backHandler={backHandler}
+                currentVersionData={currentVersionData as Topic}
+                dataProducts={currentVersionData.dataProducts}
+                deleted={currentVersionData.deleted}
+                domains={domains}
+                entityPermissions={entityPermissions}
+                isVersionLoading={isVersionLoading}
+                owners={owners}
+                slashedTopicName={slashedEntityName}
+                tier={tier as TagLabel}
+                version={version}
+                versionHandler={versionHandler}
+                versionList={versionList}
+              />
+            )
+          : null;
       }
 
       case EntityType.DASHBOARD: {
-        return (
-          <DashboardVersion
-            backHandler={backHandler}
-            currentVersionData={currentVersionData as Dashboard}
-            dataProducts={currentVersionData.dataProducts}
-            deleted={currentVersionData.deleted}
-            domains={domains}
-            entityPermissions={entityPermissions}
-            isVersionLoading={isVersionLoading}
-            owners={owners}
-            slashedDashboardName={slashedEntityName}
-            tier={tier as TagLabel}
-            version={version}
-            versionHandler={versionHandler}
-            versionList={versionList}
-          />
-        );
+        return DashboardVersion
+          ? wrapSuspense(
+              <DashboardVersion
+                backHandler={backHandler}
+                currentVersionData={currentVersionData as Dashboard}
+                dataProducts={currentVersionData.dataProducts}
+                deleted={currentVersionData.deleted}
+                domains={domains}
+                entityPermissions={entityPermissions}
+                isVersionLoading={isVersionLoading}
+                owners={owners}
+                slashedDashboardName={slashedEntityName}
+                tier={tier as TagLabel}
+                version={version}
+                versionHandler={versionHandler}
+                versionList={versionList}
+              />
+            )
+          : null;
       }
 
       case EntityType.PIPELINE: {
-        return (
-          <PipelineVersion
-            backHandler={backHandler}
-            currentVersionData={currentVersionData as Pipeline}
-            dataProducts={currentVersionData.dataProducts}
-            deleted={currentVersionData.deleted}
-            domains={domains}
-            entityPermissions={entityPermissions}
-            isVersionLoading={isVersionLoading}
-            owners={owners}
-            slashedPipelineName={slashedEntityName}
-            tier={tier as TagLabel}
-            version={version}
-            versionHandler={versionHandler}
-            versionList={versionList}
-          />
-        );
+        return PipelineVersion
+          ? wrapSuspense(
+              <PipelineVersion
+                backHandler={backHandler}
+                currentVersionData={currentVersionData as Pipeline}
+                dataProducts={currentVersionData.dataProducts}
+                deleted={currentVersionData.deleted}
+                domains={domains}
+                entityPermissions={entityPermissions}
+                isVersionLoading={isVersionLoading}
+                owners={owners}
+                slashedPipelineName={slashedEntityName}
+                tier={tier as TagLabel}
+                version={version}
+                versionHandler={versionHandler}
+                versionList={versionList}
+              />
+            )
+          : null;
       }
 
       case EntityType.MLMODEL: {
-        return (
-          <MlModelVersion
-            backHandler={backHandler}
-            currentVersionData={currentVersionData as Mlmodel}
-            dataProducts={currentVersionData.dataProducts}
-            deleted={currentVersionData.deleted}
-            domains={domains}
-            entityPermissions={entityPermissions}
-            isVersionLoading={isVersionLoading}
-            owners={owners}
-            slashedMlModelName={slashedEntityName}
-            tier={tier as TagLabel}
-            version={version}
-            versionHandler={versionHandler}
-            versionList={versionList}
-          />
-        );
+        return MlModelVersion
+          ? wrapSuspense(
+              <MlModelVersion
+                backHandler={backHandler}
+                currentVersionData={currentVersionData as Mlmodel}
+                dataProducts={currentVersionData.dataProducts}
+                deleted={currentVersionData.deleted}
+                domains={domains}
+                entityPermissions={entityPermissions}
+                isVersionLoading={isVersionLoading}
+                owners={owners}
+                slashedMlModelName={slashedEntityName}
+                tier={tier as TagLabel}
+                version={version}
+                versionHandler={versionHandler}
+                versionList={versionList}
+              />
+            )
+          : null;
       }
       case EntityType.CONTAINER: {
-        return (
-          <ContainerVersion
-            backHandler={backHandler}
-            breadCrumbList={slashedEntityName}
-            currentVersionData={currentVersionData as Container}
-            dataProducts={currentVersionData.dataProducts}
-            deleted={currentVersionData.deleted}
-            domains={domains}
-            entityPermissions={entityPermissions}
-            isVersionLoading={isVersionLoading}
-            owners={owners}
-            tier={tier as TagLabel}
-            version={version}
-            versionHandler={versionHandler}
-            versionList={versionList}
-          />
-        );
+        return ContainerVersion
+          ? wrapSuspense(
+              <ContainerVersion
+                backHandler={backHandler}
+                breadCrumbList={slashedEntityName}
+                currentVersionData={currentVersionData as Container}
+                dataProducts={currentVersionData.dataProducts}
+                deleted={currentVersionData.deleted}
+                domains={domains}
+                entityPermissions={entityPermissions}
+                isVersionLoading={isVersionLoading}
+                owners={owners}
+                tier={tier as TagLabel}
+                version={version}
+                versionHandler={versionHandler}
+                versionList={versionList}
+              />
+            )
+          : null;
       }
       case EntityType.SEARCH_INDEX: {
-        return (
-          <SearchIndexVersion
-            backHandler={backHandler}
-            breadCrumbList={slashedEntityName}
-            currentVersionData={currentVersionData as SearchIndex}
-            dataProducts={currentVersionData.dataProducts}
-            deleted={currentVersionData.deleted}
-            domains={domains}
-            entityPermissions={entityPermissions}
-            isVersionLoading={isVersionLoading}
-            owners={owners}
-            tier={tier as TagLabel}
-            version={version}
-            versionHandler={versionHandler}
-            versionList={versionList}
-          />
-        );
+        return SearchIndexVersion
+          ? wrapSuspense(
+              <SearchIndexVersion
+                backHandler={backHandler}
+                breadCrumbList={slashedEntityName}
+                currentVersionData={currentVersionData as SearchIndex}
+                dataProducts={currentVersionData.dataProducts}
+                deleted={currentVersionData.deleted}
+                domains={domains}
+                entityPermissions={entityPermissions}
+                isVersionLoading={isVersionLoading}
+                owners={owners}
+                tier={tier as TagLabel}
+                version={version}
+                versionHandler={versionHandler}
+                versionList={versionList}
+              />
+            )
+          : null;
       }
 
       case EntityType.DASHBOARD_DATA_MODEL: {
-        return (
-          <DataModelVersion
-            backHandler={backHandler}
-            currentVersionData={currentVersionData as DashboardDataModel}
-            dataProducts={currentVersionData.dataProducts}
-            deleted={currentVersionData.deleted}
-            domains={domains}
-            entityPermissions={entityPermissions}
-            isVersionLoading={isVersionLoading}
-            owners={owners}
-            slashedDataModelName={slashedEntityName}
-            tier={tier as TagLabel}
-            version={version}
-            versionHandler={versionHandler}
-            versionList={versionList}
-          />
-        );
+        return DataModelVersion
+          ? wrapSuspense(
+              <DataModelVersion
+                backHandler={backHandler}
+                currentVersionData={currentVersionData as DashboardDataModel}
+                dataProducts={currentVersionData.dataProducts}
+                deleted={currentVersionData.deleted}
+                domains={domains}
+                entityPermissions={entityPermissions}
+                isVersionLoading={isVersionLoading}
+                owners={owners}
+                slashedDataModelName={slashedEntityName}
+                tier={tier as TagLabel}
+                version={version}
+                versionHandler={versionHandler}
+                versionList={versionList}
+              />
+            )
+          : null;
       }
 
       case EntityType.STORED_PROCEDURE: {
-        return (
-          <StoredProcedureVersion
-            backHandler={backHandler}
-            currentVersionData={currentVersionData as StoredProcedure}
-            dataProducts={currentVersionData.dataProducts}
-            deleted={currentVersionData.deleted}
-            domains={domains}
-            entityPermissions={entityPermissions}
-            isVersionLoading={isVersionLoading}
-            owners={owners}
-            slashedTableName={slashedEntityName}
-            tier={tier as TagLabel}
-            version={version}
-            versionHandler={versionHandler}
-            versionList={versionList}
-          />
-        );
+        return StoredProcedureVersion
+          ? wrapSuspense(
+              <StoredProcedureVersion
+                backHandler={backHandler}
+                currentVersionData={currentVersionData as StoredProcedure}
+                dataProducts={currentVersionData.dataProducts}
+                deleted={currentVersionData.deleted}
+                domains={domains}
+                entityPermissions={entityPermissions}
+                isVersionLoading={isVersionLoading}
+                owners={owners}
+                slashedTableName={slashedEntityName}
+                tier={tier as TagLabel}
+                version={version}
+                versionHandler={versionHandler}
+                versionList={versionList}
+              />
+            )
+          : null;
       }
 
       case EntityType.API_ENDPOINT: {
-        return (
-          <APIEndpointVersion
-            backHandler={backHandler}
-            currentVersionData={currentVersionData as APIEndpoint}
-            domains={domains}
-            entityPermissions={entityPermissions}
-            isVersionLoading={isVersionLoading}
-            owners={owners}
-            slashedApiEndpointName={slashedEntityName}
-            tier={tier as TagLabel}
-            version={version}
-            versionHandler={versionHandler}
-            versionList={versionList}
-          />
-        );
+        return APIEndpointVersion
+          ? wrapSuspense(
+              <APIEndpointVersion
+                backHandler={backHandler}
+                currentVersionData={currentVersionData as APIEndpoint}
+                domains={domains}
+                entityPermissions={entityPermissions}
+                isVersionLoading={isVersionLoading}
+                owners={owners}
+                slashedApiEndpointName={slashedEntityName}
+                tier={tier as TagLabel}
+                version={version}
+                versionHandler={versionHandler}
+                versionList={versionList}
+              />
+            )
+          : null;
       }
       case EntityType.METRIC: {
-        return (
-          <MetricVersion
-            backHandler={backHandler}
-            currentVersionData={currentVersionData as Metric}
-            domains={domains}
-            entityPermissions={entityPermissions}
-            isVersionLoading={isVersionLoading}
-            owners={owners}
-            slashedMetricName={slashedEntityName}
-            tier={tier as TagLabel}
-            version={version}
-            versionHandler={versionHandler}
-            versionList={versionList}
-          />
-        );
+        return MetricVersion
+          ? wrapSuspense(
+              <MetricVersion
+                backHandler={backHandler}
+                currentVersionData={currentVersionData as Metric}
+                domains={domains}
+                entityPermissions={entityPermissions}
+                isVersionLoading={isVersionLoading}
+                owners={owners}
+                slashedMetricName={slashedEntityName}
+                tier={tier as TagLabel}
+                version={version}
+                versionHandler={versionHandler}
+                versionList={versionList}
+              />
+            )
+          : null;
       }
       case EntityType.CHART: {
-        return (
-          <ChartVersion
-            backHandler={backHandler}
-            currentVersionData={currentVersionData as Chart}
-            dataProducts={currentVersionData.dataProducts}
-            deleted={currentVersionData.deleted}
-            domains={domains}
-            entityPermissions={entityPermissions}
-            isVersionLoading={isVersionLoading}
-            owners={owners}
-            slashedChartName={slashedEntityName as unknown as string[]}
-            tier={tier as TagLabel}
-            version={version}
-            versionHandler={versionHandler}
-            versionList={versionList}
-          />
-        );
+        return ChartVersion
+          ? wrapSuspense(
+              <ChartVersion
+                backHandler={backHandler}
+                currentVersionData={currentVersionData as Chart}
+                dataProducts={currentVersionData.dataProducts}
+                deleted={currentVersionData.deleted}
+                domains={domains}
+                entityPermissions={entityPermissions}
+                isVersionLoading={isVersionLoading}
+                owners={owners}
+                slashedChartName={slashedEntityName as unknown as string[]}
+                tier={tier as TagLabel}
+                version={version}
+                versionHandler={versionHandler}
+                versionList={versionList}
+              />
+            )
+          : null;
       }
       case EntityType.DIRECTORY: {
-        return (
-          <DirectoryVersion
-            backHandler={backHandler}
-            breadCrumbList={slashedEntityName}
-            currentVersionData={currentVersionData as Directory}
-            dataProducts={currentVersionData.dataProducts}
-            deleted={currentVersionData.deleted}
-            domains={domains}
-            entityPermissions={entityPermissions}
-            isVersionLoading={isVersionLoading}
-            owners={owners}
-            tier={tier as TagLabel}
-            version={version}
-            versionHandler={versionHandler}
-            versionList={versionList}
-          />
-        );
+        return DirectoryVersion
+          ? wrapSuspense(
+              <DirectoryVersion
+                backHandler={backHandler}
+                breadCrumbList={slashedEntityName}
+                currentVersionData={currentVersionData as Directory}
+                dataProducts={currentVersionData.dataProducts}
+                deleted={currentVersionData.deleted}
+                domains={domains}
+                entityPermissions={entityPermissions}
+                isVersionLoading={isVersionLoading}
+                owners={owners}
+                tier={tier as TagLabel}
+                version={version}
+                versionHandler={versionHandler}
+                versionList={versionList}
+              />
+            )
+          : null;
       }
       case EntityType.FILE: {
-        return (
-          <FileVersion
-            backHandler={backHandler}
-            breadCrumbList={slashedEntityName}
-            currentVersionData={currentVersionData as File}
-            dataProducts={currentVersionData.dataProducts}
-            deleted={currentVersionData.deleted}
-            domains={domains}
-            entityPermissions={entityPermissions}
-            isVersionLoading={isVersionLoading}
-            owners={owners}
-            tier={tier as TagLabel}
-            version={version}
-            versionHandler={versionHandler}
-            versionList={versionList}
-          />
-        );
+        return FileVersion
+          ? wrapSuspense(
+              <FileVersion
+                backHandler={backHandler}
+                breadCrumbList={slashedEntityName}
+                currentVersionData={currentVersionData as File}
+                dataProducts={currentVersionData.dataProducts}
+                deleted={currentVersionData.deleted}
+                domains={domains}
+                entityPermissions={entityPermissions}
+                isVersionLoading={isVersionLoading}
+                owners={owners}
+                tier={tier as TagLabel}
+                version={version}
+                versionHandler={versionHandler}
+                versionList={versionList}
+              />
+            )
+          : null;
       }
       case EntityType.SPREADSHEET: {
-        return (
-          <SpreadsheetVersion
-            backHandler={backHandler}
-            breadCrumbList={slashedEntityName}
-            currentVersionData={currentVersionData as Spreadsheet}
-            dataProducts={currentVersionData.dataProducts}
-            deleted={currentVersionData.deleted}
-            domains={domains}
-            entityPermissions={entityPermissions}
-            isVersionLoading={isVersionLoading}
-            owners={owners}
-            tier={tier as TagLabel}
-            version={version}
-            versionHandler={versionHandler}
-            versionList={versionList}
-          />
-        );
+        return SpreadsheetVersion
+          ? wrapSuspense(
+              <SpreadsheetVersion
+                backHandler={backHandler}
+                breadCrumbList={slashedEntityName}
+                currentVersionData={currentVersionData as Spreadsheet}
+                dataProducts={currentVersionData.dataProducts}
+                deleted={currentVersionData.deleted}
+                domains={domains}
+                entityPermissions={entityPermissions}
+                isVersionLoading={isVersionLoading}
+                owners={owners}
+                tier={tier as TagLabel}
+                version={version}
+                versionHandler={versionHandler}
+                versionList={versionList}
+              />
+            )
+          : null;
       }
       case EntityType.WORKSHEET: {
-        return (
-          <WorksheetVersion
-            backHandler={backHandler}
-            breadCrumbList={slashedEntityName}
-            currentVersionData={currentVersionData as Worksheet}
-            dataProducts={currentVersionData.dataProducts}
-            deleted={currentVersionData.deleted}
-            domains={domains}
-            entityPermissions={entityPermissions}
-            isVersionLoading={isVersionLoading}
-            owners={owners}
-            tier={tier as TagLabel}
-            version={version}
-            versionHandler={versionHandler}
-            versionList={versionList}
-          />
-        );
+        return WorksheetVersion
+          ? wrapSuspense(
+              <WorksheetVersion
+                backHandler={backHandler}
+                breadCrumbList={slashedEntityName}
+                currentVersionData={currentVersionData as Worksheet}
+                dataProducts={currentVersionData.dataProducts}
+                deleted={currentVersionData.deleted}
+                domains={domains}
+                entityPermissions={entityPermissions}
+                isVersionLoading={isVersionLoading}
+                owners={owners}
+                tier={tier as TagLabel}
+                version={version}
+                versionHandler={versionHandler}
+                versionList={versionList}
+              />
+            )
+          : null;
       }
 
       case EntityType.DATABASE: {
-        return <DatabaseVersionPage />;
+        return DatabaseVersionPage
+          ? wrapSuspense(<DatabaseVersionPage />)
+          : null;
       }
 
       case EntityType.DATABASE_SCHEMA: {
-        return <DatabaseSchemaVersionPage />;
+        return DatabaseSchemaVersionPage
+          ? wrapSuspense(<DatabaseSchemaVersionPage />)
+          : null;
       }
 
       case EntityType.DATA_PRODUCT: {
-        return <DataProductsPage />;
+        return DataProductsPage ? wrapSuspense(<DataProductsPage />) : null;
       }
 
       case EntityType.API_COLLECTION: {
-        return <APICollectionVersionPage />;
+        return APICollectionVersionPage
+          ? wrapSuspense(<APICollectionVersionPage />)
+          : null;
       }
 
-      default:
-        VersionPage = entityUtilClassBase.getEntityDetailComponent(entityType);
+      default: {
+        const VersionPage =
+          entityVersionClassBase.getEntityDetailComponent(entityType);
 
-        return VersionPage && <VersionPage />;
+        return VersionPage ? <VersionPage /> : null;
+      }
     }
   };
