+++ openmetadata-ui/src/main/resources/ui/src/utils/PipelineVersionUtils.ts
@@ -24,10 +24,10 @@ import {
   getChangedEntityNewValue,
   getChangedEntityOldValue,
   getDiffByFieldName,
-  getTagsDiff,
-  getTextDiff,
   isEndsWithField,
-} from './EntityVersionUtils';
+} from './EntityDiffPureUtils';
+import { getTextDiff } from './EntityDiffUtils';
+import { getTagsDiff } from './EntityVersionUtilsPure';
 import { TagLabelWithStatus } from './EntityVersionUtils.interface';
 
 const handleTaskDescriptionChangeDiff = (
